Warnings fix

+++ coding/coding_tests/file_container_test.cpp
@@ -294,7 +294,7 @@ UNIT_TEST(FilesMappingContainer_Smoke)
       FileWriter w = writer.GetWriter(key[i]);
       for (uint32_t j = 0; j < count; ++j)
       {
-        uint32_t v = j + i;
+        uint32_t v = j + static_cast<uint32_t>(i);
         w.Write(&v, sizeof(v));
       }
     }
