+++ geometry/spline.hpp
@@ -28,7 +28,7 @@ public:
     double GetLength() const;
     double GetFullLength() const;
 
-    int GetIndex() const;
+    size_t GetIndex() const;
 
   private:
     friend class Spline;
@@ -40,7 +40,7 @@ public:
   private:
     bool m_checker;
     Spline const * m_spl;
-    int m_index;
+    size_t m_index;
     double m_dist;
   };
 
@@ -64,7 +64,7 @@ public:
 
     f(begin.m_pos);
 
-    for (int i = begin.GetIndex() + 1; i <= end.GetIndex(); ++i)
+    for (size_t i = begin.GetIndex() + 1; i <= end.GetIndex(); ++i)
       f(m_position[i]);
 
     f(end.m_pos);
